Added a base bootstrap repo and index.html
I’m not super well-versed in bootstrap, so if there’s a better way of
going about this please let me know and I will change it.

The index.html is a basic start page with a navbar on top with a few
links. I think it’s a good place to start and build up from.